# Conflicts:
#	Cloudents.Core/DTOs/StudyRoomDto.cs
#	Cloudents.Core/DTOs/Users/UserDto.cs
#	Cloudents.Core/Entities/ChatRoom.cs
#	Cloudents.Core/Entities/StudyRoomSessionUser.cs
#	Cloudents.Infrastructure.Data.Test/IntegrationTests/UnitOfWorkValidation.cs
#	Cloudents.Persistance/Maps/TutorMap.cs
#	Cloudents.Query/Tutor/StudyRoomQuery.cs
#	ConsoleApp/Program.cs

+++ Cloudents.Command/CommandHandler/ApplyCouponCommandHandler.cs
@@ -41,7 +41,7 @@ namespace Cloudents.Command.CommandHandler
 
             var tutorPrice = (double)tutor.Price.Price;
 
-            message.NewPrice = Coupon.CalculatePrice(coupon.CouponType, tutorPrice, (double)coupon.Value);
+            message.NewPrice = Coupon.CalculatePrice(coupon.CouponType, tutorPrice, coupon.Value);
 
         }
     }
