+++ Cloudents.Query/Admin/PaymentBySessionIdQuery.cs
@@ -82,7 +82,7 @@ namespace Cloudents.Query.Admin
                     return result;
                 }
 
-                result.StudentPayPerHour = Coupon.CalculatePrice(result.CouponType.Value, result.TutorPricePerHour, (double)result.CouponValue.GetValueOrDefault());
+                result.StudentPayPerHour = Coupon.CalculatePrice(result.CouponType.Value, result.TutorPricePerHour, result.CouponValue.GetValueOrDefault());
 
                 if (result.CouponTutor is null)
                 {
